Runwal 25 Hour Life Residence In Thane
Runwal 25 Hour Life is a premium residential project located situated next to R-Mall, Manpada, Ghodbunder Road in Thane. The project is designed to provide a luxurious lifestyle with all the modern amenities and facilities that residents could desire. The project is spread across 21 acres of land and offers exclusively designed 2 BHK and 3 BHK apartments.

The Apartments at Runwal 25 Hour Life are designed to offer a comfortable living experience with spacious rooms, large balconies, and modern fittings and fixtures. The apartments range in size from 550sq. ft. to 950+ sq. ft. and are designed to cater to the needs of both nuclear and joint families. The project offers 50+ modern amenities and facilities that include a swimming pool, gymnasium, clubhouse, landscaped gardens, and kid's play area. The project also has 24*7 hour security and power backup to ensure the safety and convenience of residents.

The Location of Runwal 25 Hour Life is another major advantage. The project is located just 1 minutes away from R-Mall, one of the largest malls in Thane. It is also located close to several reputed schools such as hospitals, and other civic amenities (which are listed below) making it an ideal choice for families.

The University of Mumbai Sub-Campus (4 mins)
Hiranandani Foundation School (5 mins)
C.P. Goenka International School (12 mins)
D.A.V. Public New Generation School (13 mins)
 
-R Mall (1 min)
-The Walk High Street (4 mins)
-D Mart- GB Road (4 mins)
-Big Centre (5 mins)
-Korum Mall (6 mins)

-Currae Multi-Speciality Hospital (3 mins)
-Bethany Hospital (6 mins)
-Life Line Hospital (7 mins)
-Hiranandani Hospital (12 mins)
-Jupiter Hospital (14 mins)

-Hiranandani Business Park (5 mins)
-MBC Infotech Park(7 mins)
-G Corp Tech Park (7 mins)
-Kalpataru Prime (8 mins)

In conclusion, Runwal 25HourLife is an excellent choice for those looking for a luxurious and comfortable living experience in Thane. The project offers a range of modern amenities and facilities, spacious apartments, and a prime location, making it an ideal choice for families and investors alike.

For more details - https://bit.ly/3IDLDy0